+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,22 +1,19 @@
 package kitchenpos.tablegroup.domain;
 
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTables;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.util.CollectionUtils;
 
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 public class TableGroup {
-    private static final int MINIMUM_TABLE_SIZE = 2;
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -24,30 +21,20 @@ public class TableGroup {
     @CreatedDate
     private LocalDateTime createdDate = LocalDateTime.now();
 
-
-    @OneToMany(mappedBy = "tableGroup")
-    private List<OrderTable> orderTables = new ArrayList<>();
+    @Embedded
+    private OrderTables orderTables;
 
     public TableGroup() {
     }
 
     public TableGroup(List<OrderTable> orderTables) {
-        validateOrderTables(orderTables);
-        this.orderTables = orderTables;
-        orderTables.forEach(orderTable -> orderTable.groupBy(this));
+        this.orderTables = new OrderTables(orderTables);
     }
 
-    public void group(List<OrderTable> orderTables) {
-        validateOrderTables(orderTables);
-        this.orderTables
-                .addAll(orderTables);
-        orderTables.forEach(orderTable -> orderTable.groupBy(this));
-    }
-
-    private void validateOrderTables(List<OrderTable> orderTables) {
-        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < MINIMUM_TABLE_SIZE) {
-            throw new IllegalArgumentException();
-        }
+    public TableGroup groupTables() {
+        orderTables.getOrderTables()
+                .forEach(orderTable -> orderTable.groupBy(this));
+        return this;
     }
 
     public Long getId() {
@@ -66,11 +53,7 @@ public class TableGroup {
         this.createdDate = createdDate;
     }
 
-    public List<OrderTable> getOrderTables() {
+    public OrderTables getOrderTables() {
         return orderTables;
-    }
-
-    public void setOrderTables(final List<OrderTable> orderTables) {
-        this.orderTables = orderTables;
     }
 }
