+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -22,7 +22,7 @@ public class TableGroupResponse {
     }
 
     public TableGroupResponse(TableGroup tableGroup) {
-        this(tableGroup.getId(), tableGroup.getCreatedDate(), tableGroup.getOrderTables()
+        this(tableGroup.getId(), tableGroup.getCreatedDate(), tableGroup.getOrderTables().getOrderTables()
                 .stream()
                 .map(OrderTableResponse::new)
                 .collect(Collectors.toList()));
